Atom Tweaks v0.1.0+c1f5afa AtomTweaks.DataCase

This module defines the setup for tests requiring access to the application's data layer.

You may define functions here to be used as helpers in your tests.

Finally, if the test case interacts with the database, it cannot be async. For this reason, every test runs inside a transaction which is reset at the beginning of the test unless the test case is marked as async.

Functions

errors_on(changeset)

Transforms changeset errors into a map of messages.

Examples

assert {:error, changeset} = Accounts.create_user(%{password: "short"})
assert "password is too short" in errors_on(changeset).password
assert %{password: ["password is too short"]} = errors_on(changeset)

errors_on(changeset, field)

Returns the list of errors on field in the changeset.

Examples

assert "password is too short" in errors_on(changeset, :password)