Atom Tweaks v0.1.0+c1f5afa AtomTweaks.Factory

Factories for generating fake database records.

Functions

build(factory_name, attrs \\ %{})

build_list(number_of_records, factory_name, attrs \\ %{})

build_pair(factory_name, attrs \\ %{})

create(_)

Specs

create(any()) :: no_return()

create(_, _)

Specs

create(any(), any()) :: no_return()

create_list(_, _, _)

Specs

create_list(any(), any(), any()) :: no_return()

create_pair(_, _)

Specs

create_pair(any(), any()) :: no_return()

factory(factory_name)

Specs

factory(any()) :: no_return()

Raises a helpful error if no factory is defined.

insert(already_built_record)

insert(factory_name, attrs \\ %{})

insert_list(number_of_records, factory_name, attrs \\ %{})

insert_pair(factory_name, attrs \\ %{})

json_params_for(key, attrs \\ [])

For when you need params for a given record to supply to an API.

note_factory()

Generates realistic-looking AtomTweaks.Releases.Note records.

params_for(factory_name, attrs \\ %{})

params_with_assocs(factory_name, attrs \\ %{})

string_params_for(factory_name, attrs \\ %{})

string_params_with_assocs(factory_name, attrs \\ %{})

token_factory()

Generates realistic-looking AtomTweaks.Accounts.Token records.

tweak_factory()

Generates realistic-looking AtomTweaks.Tweaks.Tweak records.

user_factory()

Generates realistic-looking AtomTweaks.Accounts.User records.