Atom Tweaks v0.1.0+c1f5afa AtomTweaks.Tweaks

Context for working with tweaks and their operations.

Functions

change_tweak(tweak)

Specs

change_tweak(AtomTweaks.Tweaks.Tweak.t()) :: Ecto.Changeset.t()

Creates an Ecto.Changeset for tracking tweak changes.

count_forks(tweak)

Specs

count_forks(AtomTweaks.Tweaks.Tweak.t()) :: non_neg_integer()

Counts the number of forks of the given tweak.

Note: Does not include transitive forks in this number.

create_tweak(attrs \\ %{})

Specs

create_tweak(Map.t()) ::
  {:ok, AtomTweaks.Tweaks.Tweak.t()} | {:error, Ecto.Changeset.t()}

Creates a new tweak.

fork_tweak(original_tweak, user)

Specs

fork_tweak(AtomTweaks.Tweaks.Tweak.t(), AtomTweaks.Accounts.User.t()) ::
  {:ok, AtomTweaks.Tweaks.Tweak.t()} | {:error, Ecto.Changeset.t()}

Forks the tweak by the user.

Returns the newly created tweak or an error changeset.

get_tweak!(id)

Specs

get_tweak!(binary()) :: AtomTweaks.Tweaks.Tweak.t() | no_return()

Gets a tweak by ID.

is_starred?(tweak, user)

Specs

is_starred?(
  AtomTweaks.Tweaks.Tweak.t(),
  AtomTweaks.Accounts.User.t() | nil | atom()
) :: boolean()

Checks to see if the given tweak is starred by user.

list_forks(tweak)

Specs

list_forks(AtomTweaks.Tweaks.Tweak.t()) :: [AtomTweaks.Tweaks.Tweak.t()]

Lists the forks of the given tweak.

list_stargazers(tweak)

Specs

list_stargazers(AtomTweaks.Tweaks.Tweak.t()) :: [AtomTweaks.Accounts.User.t()]

Lists the users that have starred the tweak.

list_tweaks(options \\ [])

Lists the tweaks according to the options.

Defaults to listing all original (not forked) tweaks, newest first.

Options

:for - User whose tweaks to list

:forks - When true, includes forked tweaks in the list (default: false)

:type - Includes only tweaks of the given type, if nil lists all tweaks (default: nil)