Atom Tweaks v0.1.0+c1f5afa AtomTweaks.Tweaks.Tweak

Represents a tweak.

Fields

code - Source code of the tweak

description - Markdown description of what the tweak does

title - Title of the tweak

type - The type of the tweak

Associations

Must be preloaded before they can be used.

forked_from - The AtomTweaks.Tweaks.Tweak this tweak was forked from

forks - The AtomTweaks.Tweaks.Tweak records that have been forked from this tweak

stargazers - The AtomTweaks.Accounts.User records that have starred this tweak

user - The AtomTweaks.Accounts.User that created or forked this tweak

Types

t()

Specs

t() :: %AtomTweaks.Tweaks.Tweak{
  __meta__: term(),
  code: term(),
  created_by: term(),
  description: term(),
  forked_from: term(),
  forks: term(),
  id: term(),
  inserted_at: term(),
  parent: term(),
  stargazers: term(),
  title: term(),
  type: term(),
  updated_at: term(),
  user: term()
}

Functions

filter_by_type(query, type)

Filters query to include only tweaks of type.

If nil is given for the type, the query is not filtered. This allows for easily building the query in a pipeline.

filter_by_user(query, user)

Filters query to include only tweaks that were created by user.

If nil is given for the user, the query is not filtered. This allows for easily building the query in a pipeline.

fork_params(tweak, user)

include_forks(query, arg2)

validate_fork_by_different_user(changeset, original_tweak)

Specs

validate_fork_by_different_user(Ecto.Changeset.t(), t() | binary()) ::
  Ecto.Changeset.t()

Validates that the person forking the tweak is different from the original author of the tweak.