Atom Tweaks v0.1.0+c1f5afa AtomTweaksWeb.ForkController

Handles requests for forked tweak records.

Functions

create(conn, map)

Forks a tweak for the currently logged in user.

Redirects to the newly created tweak upon success.

index(conn, map)

Displays the list of forks of a given tweak.