Atom Tweaks v0.1.0+c1f5afa AtomTweaksWeb.HerokuMetadata

A Plug that adds Heroku metadata to the metadata to be rendered on the page via the PageMetadata system.

Functions

call(conn, options)

Adds the configured Heroku metadata to the page.

get(params \\ [])

init(options)

Initialize the plug.

When given no options, the full set of Heroku environment variables are applied as metadata:

HEROKU_APP_ID

HEROKU_APP_NAME

HEROKU_DYNO_ID

HEROKU_RELEASE_CREATED_AT

HEROKU_RELEASE_VERSION

HEROKU_SLUG_COMMIT

HEROKU_SLUG_DESCRIPTION

See the Heroku documentation for what each one means.

Options

The options are mutually exclusive:

:except - Will include all environment variables except those listed

:only - Will only include the listed environment variables