Atom Tweaks v0.1.0+c1f5afa AtomTweaksWeb.PageMetadata

A system for easily setting metadata for the page before it is rendered.

If you assign some metadata to the page before it is rendered:

iex> PageMetadata.add(conn, foo: "bar")
iex> hd(conn.assigns.page_metadata)
[foo: "bar"]

And add the PageMetadata.render/1 call to the head section of the layout template:

<%= PageMetadata.render(@conn) %>

It will show up in the rendered page:

<html>
  <head>
    <meta foo="bar">
  </head>
  <body>
  </body>
</html>

Functions

add(conn, metadata)

Specs

add(Plug.Conn.t(), AtomTweaksWeb.PageMetadata.Metadata.t() | nil) ::
  Plug.Conn.t()

Adds metadata to the page.

The metadata for any type that implements the AtomTweaksWeb.PageMetadata.Metadata protocol can be added to a page.

render(conn)

Specs

render(Plug.Conn.t()) :: Phoenix.HTML.safe() | nil

Renders the metadata for the page, if it was set.

It renders each item of metadata as an individual meta tag. This function should be called in the head section of the page layout template, typically app_name_web/templates/layout/app.html.eex.