Atom Tweaks v0.1.0+c1f5afa AtomTweaksWeb.PageMetadata.Metadata protocol

Defines the metadata protocol.

To define an implementation for this protocol for a type, one must return a representation of that type as an item or list of metadata. A single item of metadata is a keyword list of attribute/value pairs intended to be rendered on an HTML meta tag. Multiple items of metadata can be returned as a list of keyword lists.

Examples

The metadatum:

[property: "og:title", content: "Some title"]

becomes the HTML:

<meta property="og:title" content="Some title" />

The metadata:

[
  [property: "og:title", content: "Some title"],
  [property: "og:description", content: "A much longer description"]
]

becomes the HTML:

<meta property="og:title" content="Some title" />
<meta property="og:description" content="A much longer description" />

Types

t()

Specs

t() :: term()

Functions

to_metadata(record)

Gets the page metadata for record.