Atom Tweaks v0.1.0+c1f5afa AtomTweaksWeb.PrimerHelpers

View helper functions for generating elements that work with GitHub's Primer CSS framework.

All functions can be used either within a template or composed together in code. Each function should always emit Phoenix.HTML.safe/0 objects or throw an exception.

Types

app_name()

Specs

app_name() :: atom()

The application name as an atom.

Functions

avatar(user, options \\ [])

Specs

avatar(AtomTweaks.Accounts.User.t(), Keword.t()) :: Phoenix.HTML.safe()

Renders the avatar element for the user.

See:Avatar element documentation

Options

:size -- value in pixels to use for both the width and height of the avatar image

code_with_heart()

Specs

code_with_heart() :: Phoenix.HTML.safe()

Renders the GitHub-style <> with ♥ by [author link] footer item.

Retrieves the author's name and URL from the application configuration for the default application for the current module. See code_with_heart/2 for more information.

code_with_heart(app_name, options \\ [])

Specs

code_with_heart(atom(), Keyword.t()) :: Phoenix.HTML.safe()

Renders the GitHub-style <> with ♥ by [author link] footer item.

Retrieves the author's name and URL from the application configuration before passing to code_with_heart/3. This information can be added to the application configuration by adding the following to your config.exs:

config :app_name,
  code_with_heart: [
    name: "Author's name",
    url: "https://example.com"
  ]

Raises a AtomTweaksWeb.PrimerHelpers.MissingConfigurationError if any of the required application configuration information is not specified and this function is called.

If passed two strings instead of an atom and a keyword list, this function will assume that you mean to call code_with_heart/3 with no options and do so for you.

code_with_heart(name, url, options)

Specs

code_with_heart(String.t(), String.t(), Keyword.t()) :: Phoenix.HTML.safe()

Renders the GitHub-style <> with ♥ by [author link] footer item.

The text in this element is intentionally left untranslated because the form of the element is intended to be recognizable in its specific format.

Options

All options are passed to the underlying HTML a element.

Examples

Phoenix.HTML.safe_to_string(AtomTweaksWeb.PrimerHelpers.code_with_heart("Author's Name", "https://example.com"))
#=> "<svg .../> with <svg .../> by <a href="https://example.com">Author's Name</a>"

counter(count)

Specs

counter(non_neg_integer()) :: Phoenix.HTML.safe()

Renders a Counter element.

See:Counter element documentation

github_link(options \\ [])

Specs

github_link(Keyword.t()) :: Phoenix.HTML.safe()

Renders a link to the project on GitHub.

Retrieves the project name or URL from the application configuration for the default application.

github_link(app_name_or_project, options)

Specs

github_link(app_name() | String.t(), Keyword.t()) :: Phoenix.HTML.safe()

Renders a link to the project on GitHub.

If the first parameter is an atom, it retrieves the project name or URL from the application configuration. Otherwise, the project can be either the GitHub owner/project identifier or the full GitHub URL.

This configuration information can be added to the application configuration by adding the following to your config.exs:

config :app_name,
  github_link: "owner/name"

If the configuration information is missing and the first parameter is an atom, a AtomTweaksWeb.PrimerHelpers.MissingConfigurationError is raised.

Options

All options are passed to the underlying HTML a element.

link_button(text, options \\ [])

Specs

link_button(String.t(), Keyword.t()) :: Phoenix.HTML.safe()

Renders a link that visually appears as a button.

Options

:to - the URL to link to

menu(block)

Specs

menu(Keyword.t()) :: Phoenix.HTML.safe()

Renders a menu element.

See:Menu element documentation

Example

Slime template:

= menu do
  = menu_item("Foo", "/path/to/foo", selected: true)
  = menu_item("Bar", "/path/to/bar")

generates:

<nav class="menu">
  <a class="menu-item selected" href="/path/to/foo">Foo</a>
  <a class="menu-item" href="/path/to/bar">Bar</a>
</nav>

menu_item(text, link, options \\ [])

Specs

menu_item(String.t(), String.t(), Keyword.t()) :: Phoenix.HTML.safe()

Renders a menu item element.

Options

:octicon - Renders an Octicon with the menu item

:selected - If true, renders the menu item as selected

All other options are passed through to the underlying HTML a element.

underline_nav(options \\ [], block)

Specs

underline_nav(Keyword.t(), Keyword.t()) :: Phoenix.HTML.safe()

Renders an UnderlineNav element.

The underline_nav_item/3 function is used to generate the nav items within the nav element.

See:UnderlineNav element documentation

Options

All options are passed through to the underlying HTML nav element.

Example

Slime template:

= underline_nav do
  = underline_nav_item "Foo", "/path/to/foo", selected: true
  = underline_nav_item "Bar", "/path/to/bar"

generates:

<nav class="UnderlineNav">
  <div class="UnderlineNav-body">
    <a class="UnderlineNav-item selected" href="/path/to/foo">Foo</a>
    <a class="UnderlineNav-item" href="/path/to/bar">Bar</a>
  </div>
</div>

underline_nav_item(text, link, options \\ [])

Specs

underline_nav_item(String.t(), String.t(), Keyword.t()) :: Phoenix.HTML.safe()

Renders an UnderlineNav-item element.

Options

:counter - When supplied with an integer value, renders a Counter element

:selected - When true, renders this item as selected

All other options are passed through to the underlying HTML a element.