Atom Tweaks v0.1.0+c1f5afa AtomTweaksWeb.Router

Routes requests to the website to the appropriate controller.

Functions

admin_checks(conn, _)

Plug pipeline of additional checks for routes that require a site admin.

api(conn, _)

Plug pipeline for requests sent to the API.

assign_current_user(conn, opts)

Fetch the current user from the session and add it to conn.assigns.

This will allows access to the current user in views with @current_user.

browser(conn, _)

Plug pipeline for requests sent from a browser.

call(conn, opts)

Callback invoked by Plug on every request.

init(opts)

Callback required by Plug that initializes the router for serving web requests.