Atom Tweaks v0.1.0+c1f5afa AtomTweaksWeb.Router.Helpers

Module with named helpers generated from AtomTweaksWeb.Router.

Functions

admin_release_note_path(conn_or_endpoint, action)

admin_release_note_path(conn_or_endpoint, action, params)

admin_release_note_path(conn_or_endpoint, action, id, params)

admin_release_note_url(conn_or_endpoint, action)

admin_release_note_url(conn_or_endpoint, action, params)

admin_release_note_url(conn_or_endpoint, action, id, params)

api_release_note_path(conn_or_endpoint, action)

api_release_note_path(conn_or_endpoint, action, params)

api_release_note_url(conn_or_endpoint, action)

api_release_note_url(conn_or_endpoint, action, params)

auth_path(conn_or_endpoint, action)

auth_path(conn_or_endpoint, action, params)

auth_url(conn_or_endpoint, action)

auth_url(conn_or_endpoint, action, params)

obsolete_route_path(conn_or_endpoint, action, user_id, tweak_id)

obsolete_route_path(conn_or_endpoint, action, user_id, tweak_id, params)

obsolete_route_url(conn_or_endpoint, action, user_id, tweak_id)

obsolete_route_url(conn_or_endpoint, action, user_id, tweak_id, params)

page_path(conn_or_endpoint, action)

page_path(conn_or_endpoint, action, params)

page_url(conn_or_endpoint, action)

page_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

static_integrity(endpoint, path)

Generates an integrity hash to a static asset given its file path.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

tweak_fork_path(conn_or_endpoint, action, tweak_id)

tweak_fork_path(conn_or_endpoint, action, tweak_id, params)

tweak_fork_url(conn_or_endpoint, action, tweak_id)

tweak_fork_url(conn_or_endpoint, action, tweak_id, params)

tweak_path(conn_or_endpoint, action)

tweak_path(conn_or_endpoint, action, params)

tweak_path(conn_or_endpoint, action, id, params)

tweak_star_path(conn_or_endpoint, action, tweak_id)

tweak_star_path(conn_or_endpoint, action, tweak_id, params)

tweak_star_url(conn_or_endpoint, action, tweak_id)

tweak_star_url(conn_or_endpoint, action, tweak_id, params)

tweak_stargazer_path(conn_or_endpoint, action, tweak_id)

tweak_stargazer_path(conn_or_endpoint, action, tweak_id, params)

tweak_stargazer_url(conn_or_endpoint, action, tweak_id)

tweak_stargazer_url(conn_or_endpoint, action, tweak_id, params)

tweak_url(conn_or_endpoint, action)

tweak_url(conn_or_endpoint, action, params)

tweak_url(conn_or_endpoint, action, id, params)

url(data)

Generates the connection/endpoint base URL without any path information.

user_path(conn_or_endpoint, action, id)

user_path(conn_or_endpoint, action, id, params)

user_star_path(conn_or_endpoint, action, user_id)

user_star_path(conn_or_endpoint, action, user_id, params)

user_star_url(conn_or_endpoint, action, user_id)

user_star_url(conn_or_endpoint, action, user_id, params)

user_token_path(conn_or_endpoint, action, user_id)

user_token_path(conn_or_endpoint, action, user_id, params)

user_token_path(conn_or_endpoint, action, user_id, id, params)

user_token_url(conn_or_endpoint, action, user_id)

user_token_url(conn_or_endpoint, action, user_id, params)

user_token_url(conn_or_endpoint, action, user_id, id, params)

user_url(conn_or_endpoint, action, id)

user_url(conn_or_endpoint, action, id, params)