Atom Tweaks v0.1.0+9f17b62 AtomTweaksWeb.StarController

Handles requests of information about stars.

Functions

index(conn, params)

Shows the list of stars for the named user.

toggle(conn, params)

Toggles whether the current user has starred the tweak.

After the state has been toggled, it redirects back to wherever it came from.