Atom Tweaks v0.1.0+c1f5afa Support.Setup

Functions that are designed to be used as chained setup callbacks.

Each callback takes certain values from the context state and merges in new or updated values. Because they only get supplied the context as a parameter, each function has a contract around the keys that are used.

Categories

insert functions create database records

request functions navigate to web pages

Examples

Inserts a user, creates a tweak for that user, and then stars that tweak for that user:

setup [:insert_user, :insert_tweak, :insert_star]

Functions

fork_tweak(map)

Forks a tweak.

Inputs

:tweaks

Outputs

:fork_user -- User that forked the tweak

:fork_tweak -- New fork tweak

:forked_tweak -- Tweak that was forked

insert_init_tweak(context)

Inserts an init tweak.

Outputs

:init_tweak -- Tweak that was created

insert_release_note(context)

Inserts a release note.

Outputs

:note -- Release note that was inserted into the database

insert_site_admin(context)

Inserts a site admin user into the database.

Outputs

:user - Site admin user record

insert_star(context)

Stars a tweak by a user.

Inputs

:user - User to star the tweak

:tweak - Tweak to be starred

Outputs

:star - Star record

insert_style_tweak(context)

Inserts a style tweak.

Outputs

:style_tweak -- Tweak that was created

insert_tokens(context)

insert_tweak(context)

Inserts a single tweak for a user.

Inputs

Inserts a tweak for the first of:

:current_user

:user

Inserts a new user

Outputs

:tweak - New tweak record

:user - User record that created the tweak

insert_user(context)

Inserts a user into the database.

Outputs

:user - New user record

insert_user_with_tweaks(context)

Inserts a user with three tweaks into the database.

Outputs

:user - New user record

:tweaks - New tweak records

invalid_token_params(context)

invalid_tweak_params(context)

Creates an invalid set of parameters for a tweak.

Outputs

:tweak_params -- Invalid params

log_in(context)

Logs in as a user.

Inputs

:conn -- Connection object

Logs in as (in priority order):

:user

Inserts a new user into the database

Outputs

:current_user - User record used to log in

request_admin_release_note_edit(context)

Requests the admin release note edit page.

Inputs

:conn -- Connection object

:note -- Release note to edit

Outputs

:conn -- Updated connection object

request_admin_release_note_index(context)

Requests the admin release notes index page.

Inputs

:conn -- Connection object

Outputs

:conn -- Updated connection object

request_admin_release_note_new(context)

Requests the admin release note new page.

Inputs

:conn -- Connection object

Outputs

:conn -- Updated connection object

request_admin_release_note_show(context)

Requests the admin release note show page.

Inputs

:conn -- Connection object

:note -- Release note to show

Outputs

:conn -- Updated connection object

request_create_tweak(context)

Requests the create tweak page.

Inputs

:conn -- Connection object

:current_user - Currently logged in user record, if any

:request_user - User record requesting the page

:tweak_params - Parameters to use to create the tweak

Outputs

:conn - Updated connection object

request_edit_tweak(context)

Requests the edit tweak page.

Inputs

:conn - Plug.Conn object

:tweak - Tweak record to edit

Outputs

:conn - Plug.Conn object after the page is rendered

request_forks(context)

request_new_tweak(context)

Requests the new tweak page.

Inputs

:conn - Plug.Conn object

Outputs

:conn - Plug.Conn object after the page is rendered

request_page_about(context)

Requests the about page.

request_page_index(context)

Requests the home page.

request_page_release_notes(context)

Requests the release notes page.

request_show_tweak(context)

Requests the show tweak page.

Inputs

:conn - Plug.Conn object

:tweak - Tweak record to show

Outputs

:conn - Plug.Conn object after the page is rendered

request_show_user(context)

Requests the show user page.

Inputs

:conn - Plug.Conn object

:current_user - Currently logged in user record

:request_user - User record to show

:user - User record to show

Outputs

:conn - Plug.Conn object after the page is rendered

:path - Path that was navigated to

request_stars(context)

Navigates to the stars page for a user.

Inputs

:conn - Plug.Conn object

:user - User whose stars we will view

Outputs

:conn - Updated Plug.Conn object

:path - Path that was navigated to

request_user(context)

Creates a new user in the database to request a page.

Outputs

:request_user - User record that requests a page

request_user_token_create(context)

request_user_token_index(context)

Requests the user's token index page.

request_user_token_new(context)

Requests the user's new token page.

valid_token_params(context)

Generates valid token parameters.

valid_tweak_params(context)

Generates valid tweak parameters.